refactor: extract coverage state into separate Zustand store (#159)
Move coverageSamples, isSimulationRecomputing, simulationProgress,
simulationRunToken, and recomputeCoverage action out of appStore.ts
into a dedicated coverageStore.ts. Components subscribe to coverage
state via useCoverageStore, narrowing the re-render fan-out during
coverage recompute lifecycle.
